Add event series to My Stuff view.
