+++ app/models/event_series.rb
@@ -1,5 +1,6 @@
 class EventSeries < ApplicationRecord
   include Delegable
+  include MultiEventable
   extend FriendlyId
 
   enum scoring_method: [:time, :rank, :points]
